# backend/src/main/java/com/interviewme/controller/AuthController.java
package com.interviewme.controller;

import com.interviewme.common.dto.AuthResponse;
import com.interviewme.common.dto.LoginRequest;
import com.interviewme.common.dto.RegisterRequest;
import com.interviewme.common.dto.UserInfoResponse;
import com.interviewme.service.AuthService;
import jakarta.validation.Valid;
import lombok.RequiredArgsConstructor;
import lombok.extern.slf4j.Slf4j;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.security.core.Authentication;
import org.springframework.web.bind.annotation.*;

@RestController
@RequestMapping("/api/auth")
@RequiredArgsConstructor
@Slf4j
public class AuthController {

    private final AuthService authService;

    @PostMapping("/register")
    public ResponseEntity<AuthResponse> register(@Valid @RequestBody RegisterRequest request) {
        log.info("Registration request for email: {}", request.email());
        AuthResponse response = authService.register(request);
        return ResponseEntity.status(HttpStatus.CREATED).body(response);
    }

    @PostMapping("/login")
    public ResponseEntity<AuthResponse> login(@Valid @RequestBody LoginRequest request) {
        log.info("Login request for email: {}", request.email());
        AuthResponse response = authService.login(request);
        return ResponseEntity.ok(response);
    }

    @GetMapping("/me")
    public ResponseEntity<UserInfoResponse> getCurrentUser(Authentication authentication) {
        log.info("Get current user request");
        String email = authentication.getName();
        UserInfoResponse response = authService.getCurrentUser(email);
        return ResponseEntity.ok(response);
    }
}